pyproject.toml
[project]
name = "micro-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "harish gehlot", email = "gehloth03@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "docling>=2.91.0",
    "hydra-core>=1.3.2",
    "langchain-huggingface>=1.2.2",
    "omegaconf>=2.3.0",
    "prek>=0.3.11",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "rust-just>=1.50.0",
    "sentence-transformers>=5.4.1",
    "transformers>=5.6.2",
]

[project.scripts]
index = "micro_llm.entrypoints.indexing_endpoint:main"

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"
